requirements.txt
absl-py==0.13.0
anyio==3.3.0
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.3
cloudpickle==1.6.0
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cycler==0.10.0
d2l==0.16.7
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
dm-tree==0.1.6
entrypoints==0.3
et-xmlfile==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==2.0.2
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
idna==2.10
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
keras==2.7.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==12.0.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
openai==0.9.3
opencv-python==4.5.2.52
openpyxl==3.0.7
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==1.2.4
pandas-stubs==1.1.0.12
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydub==0.25.1
Pygments==2.9.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.2
selenium==4.1.0
Send2Trash==1.7.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
tensorflow-probability==0.13.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
urllib3==1.26.6
voila==0.2.10
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.0.0
